consolidated Neo4jService tests into single file
